# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog Berita API",
    "version": "1.0.0",
    "description": "API documentation for Blog Berita application. This API provides endpoints for managing news articles, categories, tags, comments, advertisements, and user authentication."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://backend-blog-berita.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/admin/posts/{id}": {
      "put": {
        "tags": [
          "Admin Blog"
        ],
        "summary": "Update a news post",
        "description": "Update an existing blog post with optional thumbnail upload (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Post ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5,
                    "example": "Perkembangan AI di Indonesia 2026"
                  },
                  "slug": {
                    "type": "string",
                    "example": "perkembangan-ai-di-indonesia-2026"
                  },
                  "content": {
                    "type": "string",
                    "minLength": 15,
                    "example": "<p>Perkembangan kecerdasan buatan di Indonesia semakin pesat...</p>"
                  },
                  "category_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published"
                    ],
                    "example": "published"
                  },
                  "tag_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      1,
                      2,
                      3
                    ]
                  },
                  "meta_title": {
                    "type": "string",
                    "example": "AI Indonesia 2026 - Perkembangan Terbaru"
                  },
                  "meta_description": {
                    "type": "string",
                    "example": "Simak perkembangan terbaru AI di Indonesia tahun 2026"
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary"
                  },
                  "remove_thumbnail": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "News post updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Berita berhasil diperbarui"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "title": {
                          "type": "string",
                          "example": "Perkembangan AI di Indonesia 2026"
                        },
                        "slug": {
                          "type": "string",
                          "example": "perkembangan-ai-di-indonesia-2026"
                        },
                        "content": {
                          "type": "string",
                          "example": "<p>Perkembangan kecerdasan buatan di Indonesia semakin pesat...</p>"
                        },
                        "category_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "draft",
                            "published"
                          ],
                          "example": "published"
                        },
                        "thumbnail": {
                          "type": "string",
                          "nullable": true,
                          "example": "https://example.com/uploads/thumbnail.jpg"
                        },
                        "meta_title": {
                          "type": "string",
                          "nullable": true,
                          "example": "AI Indonesia 2026 - Perkembangan Terbaru"
                        },
                        "meta_description": {
                          "type": "string",
                          "nullable": true,
                          "example": "Simak perkembangan terbaru AI di Indonesia tahun 2026"
                        },
                        "created_at": {
                          "type": "string",
                          "example": "2026-06-29T10:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Post not found"
          },
          "409": {
            "description": "Slug already used"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "Admin Blog"
        ],
        "summary": "Get post by ID",
        "description": "Retrieve a single post by its ID (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Post ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Post details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Detail berita berhasil diambil"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "title": {
                          "type": "string",
                          "example": "Perkembangan AI di Indonesia 2026"
                        },
                        "slug": {
                          "type": "string",
                          "example": "perkembangan-ai-di-indonesia-2026"
                        },
                        "content": {
                          "type": "string",
                          "example": "<p>Perkembangan kecerdasan buatan di Indonesia semakin pesat...</p>"
                        },
                        "category_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "draft",
                            "published"
                          ],
                          "example": "draft"
                        },
                        "thumbnail": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        },
                        "meta_title": {
                          "type": "string",
                          "nullable": true,
                          "example": "AI Indonesia 2026 - Perkembangan Terbaru"
                        },
                        "meta_description": {
                          "type": "string",
                          "nullable": true,
                          "example": "Simak perkembangan terbaru AI di Indonesia tahun 2026"
                        },
                        "created_at": {
                          "type": "string",
                          "example": "2026-06-29T10:00:00.000Z"
                        },
                        "category": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "name": {
                              "type": "string",
                              "example": "Teknologi"
                            },
                            "slug": {
                              "type": "string",
                              "example": "teknologi"
                            }
                          }
                        },
                        "post_tags": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "example": 1
                              },
                              "post_id": {
                                "type": "integer",
                                "example": 1
                              },
                              "tag_id": {
                                "type": "integer",
                                "example": 1
                              },
                              "tag": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "integer",
                                    "example": 1
                                  },
                                  "name": {
                                    "type": "string",
                                    "example": "AI"
                                  },
                                  "slug": {
                                    "type": "string",
                                    "example": "ai"
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "example": {
                        "id": 1,
                        "user_id": 1,
                        "title": "Perkembangan AI di Indonesia 2026",
                        "slug": "perkembangan-ai-di-indonesia-2026",
                        "content": "<p>Perkembangan kecerdasan buatan di Indonesia semakin pesat...</p>",
                        "category_id": 1,
                        "status": "draft",
                        "thumbnail": null,
                        "meta_title": "AI Indonesia 2026 - Perkembangan Terbaru",
                        "meta_description": "Simak perkembangan terbaru AI di Indonesia tahun 2026",
                        "created_at": "2026-06-29T10:00:00.000Z",
                        "category": {
                          "id": 1,
                          "name": "Teknologi",
                          "slug": "teknologi"
                        },
                        "post_tags": [
                          {
                            "id": 1,
                            "post_id": 1,
                            "tag_id": 1,
                            "tag": {
                              "id": 1,
                              "name": "AI",
                              "slug": "ai"
                            }
                          },
                          {
                            "id": 2,
                            "post_id": 1,
                            "tag_id": 2,
                            "tag": {
                              "id": 2,
                              "name": "Machine Learning",
                              "slug": "machine-learning"
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin Blog"
        ],
        "summary": "Delete a news post",
        "description": "Delete a blog post by ID (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Post ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Berita berhasil dihapus"
                    },
                    "data": {
                      "type": null,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}